chore: moved class Goal and cleaned up FirebaseModel
